feat: move SVG to `icons.js`
